Get user (#14)
* Get account & get specific account

deze gets zijn gefixed

* get specific user

get specific user gefixed

* Minor changes

+++ src/main/java/io/inholland/groep4/api/UsersApi.java
@@ -38,7 +38,7 @@ public interface UsersApi {
     @RequestMapping(value = "/users/{id}",
             produces = { "application/json" },
             method = RequestMethod.GET)
-    ResponseEntity<User> getSpecificUser(@Parameter(in = ParameterIn.PATH, description = "The user ID", required=true, schema=@Schema()) @PathVariable("id") Integer id);
+    ResponseEntity<User> getSpecificUser(@Parameter(in = ParameterIn.PATH, description = "The user ID", required=true, schema=@Schema()) @PathVariable("id") Long id);
 
 
     @Operation(summary = "Get users", description = "Get users", security = {
